refactor(TECHDEBT-31): relocate trigger_services into FastAPI adapter
Move vultron/api/v2/backend/trigger_services/ into
vultron/adapters/driving/fastapi/ per architectural cleanup plan:

- Move HTTP request body models (_models.py) to
  vultron/adapters/driving/fastapi/trigger_models.py
- Move domain_error_translation() and translate_domain_errors() from
  _helpers.py into vultron/adapters/driving/fastapi/errors.py
- Merge thin adapter delegates (case.py, embargo.py, report.py) into
  new vultron/adapters/driving/fastapi/_trigger_adapter.py
- Update all three trigger routers to import from new adapter-layer
  locations (no shims)
- Update test/api/v2/backend/test_trigger_services.py imports
- Delete vultron/api/v2/backend/trigger_services/ entirely

After this change, vultron/api/v2/ contains only data/actor_io.py
(pending VCR-014) and two __init__.py stubs.

996 tests pass.
