+++ vultron/adapters/driving/fastapi/trigger_models.py
